create a nomination badge
